Backend/rios-del-desierto=backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
python-dotenv
pandas
openpyxl